+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+ - ✨ Added the ability to delete AOIs with the `delete` key
  - 🛠️ Updated track plot window to reset zoom view when axes are changed.
  - 🛠️ Updated data loader to allow loading a slice of imagery incrementally.